requirements.txt
requests
dj-database-url
Django>1.11.0,<2.0.0
django-auth-ldap ~= 1.5.0
django-contrib-comments
gunicorn
Pillow
psycopg2
reportlab
sqlparse
djangorestframework